requirements.txt
discord.py==2.2.2
duckduckgo_search==2.8.6
python-dotenv==1.0.0
Requests==2.28.2
